log instead of halting when workunit-logger encounters unserializable metadata
The workunit-logger takes all of the work units and dumps them as
json.  This *mostly* works, but occasionally a workunit has metadata
attached that is not serializable by `json.dumps`.  See #21792 for an
example of that whack-a-mole.

I think it makes more sense to treat the workunit-logger as a "best
effort" sort of system that logs and then continues instead of
imposing a "json serializable" requirement deep throughout the engine.
If we were to integrate something a more useful export format would be
OpenTelemetry anyway. In particular I don't think it makes sense to
force *json* serializability since as per
<https://stackoverflow.com/questions/3768895/how-to-make-a-class-json-serializable>
there still isn't an idiomatic Python way to do that.

(The "mole" this triggered for me was `pants.engine.addresses.Address`
being attached to `HelmChart`.)