Don't run a full CI on devcontainer changes (#22096)
The devcontainer is strictly for local development, so running 
tests/packaging is unnecessary when it changes.

To facilitate this change:
- Change the terminology "Affected" to the more general "ChangeLabel".
- Move the logic to determine whether full CI needs to run or not into
the classifier script itself, via a "no code" label. This replaces the
"docs only"
concept that was previously implemented in the CI script. Now that more
labels are in play, it is easier to implement this in Python than in
shell.