Restart docker environment container when MIA (#22128)
Closes #18209

This PR adds a `CapturedWorkdirError` enum type with 3 variations:
`Timeout`, `Retryable`, and `Fatal`. By doing this we can propagate
structured errors up the call stack.

In the case of a `DockerServerResponseError`, we return a retryable
error. If it's a server error with status code 404, we prune the old
container from the container cache.

We also add a `CommandSpawnError` specific to the docker container
`spawn` method.

For tests, we add a parameterization for most tests where 1) we start
with no container 2) we start with an existing container 3) we start
with a container in the container cache, but no existing container (i.e.
404 error).