Reorder the args of the environment_vars_subset rule. (#22217)
The EnvironmentVarsRequest arg is the one commonly provided by the
caller, with the CompleteEnvironmentVars provided implicitly.

Putting the commonly explicit arg first allows for a simpler and
expected calling idiom:

`await environment_vars_subset(EnvironmentVarsRequest(...,
**implicitly())`

With the explicit arg second in declaration order, this yields a long an
incomprehensible solver error, which ends up being fixed by

`await environment_vars_subset(**implicit(EnvironmentVarsRequest(...)))`

but that is not going to be easily knowable.

This also allows the migrate-call-by-name tool to work in more cases.