Provide selected python interpreter to non-internal pex packaging via --python-path (#22160)
This propagates Pants' interpreter selection into the PEX build when
building local-compatible `pex_binary` target, rather than have the
building process do a full-system-wide search for compatible
interpreters.

In particular, when building a pex in the "default" mode/for the local
system, without an explicit `complete_platforms` and not for
internal-use only, we now do Pants-side interpreter selection and
provide a `--python-path` argument to that chosen interpreter, to tell
Pex where to find the relevant interpreter, rather than letting Pex do
the search again.

A Pex-based search of `PATH` (and other paths) may not be able to find
an interpreter if the user is using a Python provider backend, which
will have Pants install and manage Python interpreters without being
externally installed.

Fixes #21048 

Closes #21948 (replaced by this work)

Co-authored-by: Tom Solberg <me@sbg.dev>