Include Action and Command digests in workunit metadata (#22212)
Implements https://github.com/pantsbuild/pants/issues/22203

Includes Action & Command digests in the process workunit metadata.
Using these its now easier to get the output of process executions from
a remote cache.

I've added these to the main `process` workunit which already has a
bunch of metadata on it
[execute_process.rs](https://github.com/pantsbuild/pants/blob/0374be8b3465fc0d8e69b8db34240c29945734b2/src/rust/engine/src/nodes/execute_process.rs#L231-L267)

I tried to find a way to include the process `output_directory` digest,
but that turned out to be complicated as its not materalized until its
written to the remote cache in
[make_action_result](https://github.com/pantsbuild/pants/blob/0374be8b3465fc0d8e69b8db34240c29945734b2/src/rust/engine/process_execution/remote/src/remote_cache.rs#L200-L257)
but that occurs in the background after the workunit has completed. So
I've omitted that for now unless someone has a better idea.