Add Trivy security scanner (#21780)
Trivy is a scanner for finding security vulnerabilities and
misconfigurations in Docker images, Helm charts, and Terraform modules
(among others). This MR adds it for all of those.

I made it as a single MR to minimise churn in the common code (like the
invocation of Trivy). Let me know if you'd rather I split this up.

---------

Co-authored-by: cburroughs <chris.burroughs@gmail.com>