Fix run_rule_with_mocks(). (#22273)
The original motivation was to get it to work with call-by-name 
inside concurrently().

But this change ended up being more substantial, as it fixes several
longstanding bugs in this code.

In particular, the validation of mocks was broken, but, due to another
bug, it was rarely run, so we didn't notice...

Also tweaks a handful of tests to get them to pass under the new, 
stricter (but more correct) regime.