Allow bypassing pants.backend.docker.utils.suggest_renames (#22250)
Allow users to bypass the `suggest_renames` functionality of enriching
`docker_image` failures with additional log metadata. This feature is
typically helpful for end users, but as discussed in #22246, is
currently computationally expensive.

Adding this subsystem option will provide a way for users who are facing
the aforementioned performance issue to opt-out of this feature as a
work-around.