Simplify rule mock validation. (#22321)
Instead of validating in a separate pass, validate
on the fly.

This removes a lot of clunky, repetitive and buggy code.
And since, until recently, the preliminary validation pass
wasn't invoked in practice in most cases due to various bugs,
we can be fairly sure it wasn't particularly important...

This is a preliminary change to make it easier to support
call-by-name of unions (the specific union member rule to be
invoked can only be determined on the fly).