improve performance of the PBS release scraping script (#22378)
The PBS release scraping script was slow because it iterated over all
releases even when just checking for recent releases. There is no need
to do that since releases are vended in descending order by creation
time from the most recent release.

Update the script to only scan for recent releases until any
already-scraped release is seen (which should be the most recent known
release), then stop scanning releases.

Also reduces the API pagination size to 10 so unnecessary results are
not returned (and to reduce the size of the API responses) and adds some
additional logging messages.