feat: Support TSX grammar, allowing inferring type imports in typescript (#22363)
Switch the grammar to the most permissive (TSX). While TSX is not a true
superset of TS (the `<Casting>` annotation), the likelihood of an import
statement being affected by the TS <-> TSX discrepancies are low enough
that the complexities to handle multiple grammars aren't worth it. They
are both supersets of JS regardless so this _should_ not affect users
negatively.

Infering type dependencies is likely always what you'd want, either the
runtime will require the file (node-ts), or the transpiler is going to
optimize it away in the `node_build_script`.

I think it is ok that the dependency parser is called "javascript".