remove `pkg_resources.WorkingSet` from plugin resolver and other init logic (#22377)
The `pkg_resources` module is deprecated and needs to be removed going
forward in favor of the `importlib` and `packaging` modules. The
immediate impact of keeping `pkg_resources` in the code base though is
the inability to upgrade `mypy` due to some type errors. See
https://github.com/pantsbuild/pants/pull/21873 for that attempt to
upgrade mypy.

The plugin resolver and extension loader use `pkg_resources.WorkingSet`
extensively to keep track of plugin distributions requested by the user.
There is unfortunately no equivalent to `WorkingSet` though in the
`importlib` and `packaging` modules.

One solution would be to vendor `pkg_resources.WorkingSet`. I tried this
first but found the process to be unwieldly as a lot of logic from
`pkg_resources` had to be included to make `WorkingSet` actually work,
and I still kept finding more logic to include to make vendoring work.

Instead, this PR removes `WorkingSet` entirely in favor of just using
the existing `importlib.metadata` and `packaging` APIs. A key point is
that `PluginResolver.resolve_plugins` _already_ mutably modified
`sys.path` even while using `WorkingSet`. Thus, it was simpler to retain
the `sys.path` modifications and get rid of `WorkingSet` and have tests
rely on the `sys.path` modification occurring.

The relevant tests also use a context manager to save the original
`sys.path` and restore it when the test is done. This solves the issue
of tests needing to pass in a custom `WorkingSet`.