Support for polymorphic rules (#22375)
An `@rule` can be annotated with `polymorphic=True`, in which
case, if it has a param of a union base type B, then calling it by
name with an instance of M which is a member type of B, 
will dispatch to a `@rule` with the same output type
and a param of type M.

This is the call-by-name equivalent of union Gets.

The "vtable" (mapping from member type to rule 
implementation for that member type) is computed automatically.
This has the advantage that those existing rules, currently
called via union Get, do not need to be modified or annotated.

There may however need to be small changes to rule structure
to ensure that there is a matching rule for each union member.
You can see an example in this PR.

This PR applies this functionality to three very different cases:
PackageFieldSet, GenerateSourcesRequest and GenerateTargetsRequest.