Move most of the remaining tests from `tests/` to `src/`. (#22439)
Our style is to have tests in the `src/` dir alongside the code they
test. There were a few tests remaining in `tests/` and this moves
almost all of them, renaming them from test_foo.py to foo_test.py
as needed.

There are still a few integration tests in the old location. They will
be moved in a future change.