refactor environments rules to the separate namespace (#22438)
Target types for environments were previously refactored in #22352 to a
separate `pants.core.environments` namespace. This PR finishes the job
and moves the related rules to that namespace as well from
`pants.core.util_rules.environments`. (The original refactor was
necessary to remove an import cycle encountered with the call-by-name
migration.)