Remove the never-used SOAP backend. (#22447)
This code was originally merged in #14545, but never 
had a corresponding `register.py` (see that PR for why)
and was therefore never actually activated or used.

The author of that code, @alonsodomin, wrote it to 
support a legacy codebase, and later upstreamed it, 
but they now say:

"I don't think this has any reason right now to be in the repo. 
SOAP is very niche and phased out in most places"

Again, this could never have been used as-is by any user
since there is no associated register.py. So we have
no idea if this code works, or any interested users.
But keeping it is not free (for example, should
we port it to call-by-name?) so since there is no benefit,
we should eliminate the cost.