Fix polymorphic call-by-name (#22482)
The original implementation in #22375 incorrectly assumed
that call-by-name could ignore the `in_scope_types` 
of a `@union`, and instead compute the DependencyKey using
the types in the polymorphic rule signature, and just pass
an empty `in_scope_types` to the DependencyKey, to indicate
a reentry. 

However in some unusual cases where the in_scope_type isn't
explicitly provided, this causes the solver to fail. 

This change restores correct `in_scope_types` behavior.

There could be a way to get this to work, and it would be more
intuitive to do so (since it would more closely mirror the behavior
of virtual method signatures), but it would require more changes
that we want to make right now. So we can revisit this when we
are fully call-by-name.