JVM: Fix coursier classifier conflicts (#22582)
When getting lockfile entries for coordinates, classifier was not taken
into account. So for example if an artifact is published with the same
coordinates but with different classifiers, the order in the lockfile
determined which one was actually pulled in to classpath.

I've added a test for a real life example where the testlib-classified
jar of gax-grpc is pulled in before this fix, which would break at
runtime. The workaround is quite cumbersome since you have to exclude
the affected library from anything that directly or transitively depends
on it, then add it manually. It can also be tricky to debug what is
going on.