Update default interpreter constraints for tools and integration tests (#22578)
And update default lockfiles to match. 

We now support `>=3.9,<3.14` by default.

Many tools (including pip) are dropping 3.8 support, so we should
stop trying to use it. And accepting anything `<4` by default 
was never a good idea.

The test change only affects this repo.

The tool change might affect users, if they happened to be relying
on running tools on 3.8 specifically, but the solution is to regenerate
a custom lockfile, or ensure the presence of a 3.9+ interpreter.