plumb through Pex's exclude & override flags to the resolve subsystem (#22607)
Pex has long supported the `--exclude` and `--override` as sharp tools
to manipulate the requirements of transitive dependencies. For example,
to remove an upper bound or avoid pulling in a large dependency whose
codepaths you are confident will not be used.

This plumbs them through to the Python setup subsystem following the
existing patterns.

NOTE: `--override` was introduced in Pex
[2.8.0](https://github.com/pex-tool/pex/releases/tag/v2.8.0), so the
current Pex `version_constraints` of `>=2.13.0,<3.0` do not need to be
updated.

---------

Co-authored-by: Benjy Weinberger <benjyw@gmail.com>