rename `protos::gen` to `protos::pb` in advance of Rust 2024 Edition (#22635)
In Rust 2024 Edition, `gen` is now a (reserved) keyword which means that
our crates would need to use raw identifier syntax to refer to
`protos::gen` which is the root of our generated protobuf Rust code.
Let's make life easier for our eventual upgrade to Rust 2024, and just
rename `gen` to something which is not a keyword in Rust 2024 Edition,
namely `protos::pb`.