Support mutually recursive call-by-name rules. (#22648)
The issue is that the `@rule` decorator eagerly analyzes
the awaitables used within the rule function's body, 
and expects those to already be defined. This is
not possible with mutual recursion, due to circularity.

This change introduces a `RuleDescriptor` that can be
inferred statically from the AST, without the rule being
defined, and contains just enough information for the 
awaitable analysis to work.

These are currently only supported between rules defined
at the top level of the same module (including a rule
calling itself).

We don't support recursion involving rules defined in
inner scopes. Doing so would require more substantial
changes, as a function does not have a reference to
its enclosing scope, and recreating this from 
`inspect.getclosurevars()` is more involved.