Port engine tests to call-by-name. (#22667)
These mostly test behavior that is equally relevant
for Get and for call-by-name, so we might as well
test it for the latter now.

A handful of its tests are only pertinent for Gets, 
and so are marked as such, or in some cases
deleted entirely, since they were already skipped,
and the "reachability" they ostensibly would test
for is not a thing in call-by-name.