Optionally panic if a Get is used. (#22670)
The option is controlled by an env var directly, since some rule
processing happens before options are available.

We raise an exception in the RuleVisitor if a Get is seen in 
a rule, and we also panic in the rust-side Get constructor
if a Get is created without first having been visited by
RuleVisitor (e.g., in tests).

Currently the env var is not set. We will flip it in a branch
to see what breaks.

I have manually tested that when the env var is set on
the pants client cmd line, the expected things happen
if a Get is found.