deprecate experimental rust backend (#22710)
This backend has had no meaningful development since the initial commit,
not used by any contributor, what it does could likely be replaced by
adhoc_tool and friends, and there is an out-of-tree plugin that is a
more plausible path forward.

Recent breakage has revealed that our CI setup is not currently #22704
equipped to test this plugin.

I tired to highlight in the note that this this was a good attempt and
more Rust support would also be welcome. Wordsmiths welcome.