Regenerate in-repo Python lockfiles (#22714)
The first commit is a small change to the script that 
generates the default lockfiles, and is the one worth
reviewing manually in detail.

The next 3 commits are the result of generating lockfiles:
- Custom lockfiles in our repo
- Default lockfiles we embed and ship to users
- Lockfiles used in tests

Finally the last commit is some fixes to `resource()` targets
in BUILD files, to ensure they capture the metadata of
embedded default lockfiles.