nfpm.native_libs.elfdeps: Use SOInfo dataclass
We need slightly different things for deb vs rpm. For deb, we need just
the soname to search for relevant packages, and parsing the so_info
string did not seem wise when I can just preserve the data as elfdeps
returned it. So, we now have a SOInfo dataclass (a limited mirror of the
elfdeps.SOInfo dataclass).