nfpm.native_libs.scripts: Use aiohttp-retry on flaky API
NOTE: order of client and task group in async with block is very
important. The task group needs to be the innermost block so that it can
appropriately wait for the task group to finish. Putting the client as
the innermost block closes the client before any of the tasks can
finish leading to a bunch of errors returned by the task group.