Fix failing local store tests. (#22742)
These tests started failing on some version of macOS 14, 
possibly due to differences in how the OS reports storage
of mmapped files. 

This change tries to be a bit more nuanced in computing
the expected shrinkage, which on macOS is now less
than it was previously. So this is slightly less naive, 
but could still fail after future OS updates, hence the 
comments to that effect.