[windows] Add windows support in the stdio crate. (#22745)
The `stdio` crate depended on the unix-specific RawFd.

This change removes that dependency, replacing it
with our own C file descriptor type, `CFd`.

For background: 

This crate is used for IO against the Python side's 
stdin/stdout/stderr. We pass int32 C fds (as returned by
`fileno()`) from  Python into Rust, cast those to `RawFd`,
and then wrap those in Rust `File` objects. To pass file
references back into Python we grab the `RawFd` back
from the `File`. 

This works because on unix the raw file descriptor is used
directly by the stdlib. So while we were trading in `RawFd`
what we actually meant was "C stdlib file descriptor".

However on windows the C file descriptor type *wraps*
the underlying `RawHandle`, and is not identical to it. 
It is easy to go from C file descriptor to `RawHandle`, 
but the converse is not possible.

Fortunately since all our `RawHandle`s originate from
Python-side filenos, we can store those, as `CFd` structs,
alongside the corresponding `File` objects.