A new command-line parser.
This supports the command/subcommand paradigm, rather than
the "list of goals" paradigm of the existing cli parser.

This is experimental, not wired up to anything yet, and intended
to support a "next-generation Pants" (pants_ng for short).

It is not intended to support all current Pants 2 use-cases, but
is a much simpler design based on our experience with the
complexities of the current cli parser.

In particular, this new cli parser does not need to know about
any registered options (unlike the old one, which uses them to
tell the difference between specs and goals). This parser instead
requires specs to contain a filepath separator.

This cli parser currently supports one cmd (and optionally one
subcommand) per invocation. If we want to support multiple
cmd/subcommand pairs in a single invocation in the future we
can do so by using a delimiter that isn't interpreted by the shell.

See https://github.com/pantsbuild/pants/discussions/22692 for context.