only set pex --override during lockfile generation
For many cli flags we tend towards a "belt and suspenders" approach
and set them broadly -- and perhaps redundantly -- instead of only in
the narrow case where we are sure they are needed.  This works fine in
most cases but for --override Pex has an explicit check [1] so that in
Pants terms one most override when generating a lockfile, not when
using or subsetting it.

[1] https://github.com/pex-tool/pex/blob/682536efd427b07c571c5c5718a56ff9be1edf40/pex/bin/pex.py#L1012