Fix crypto provider initialization in grpc_util tests. (#22811)
The initialization is intended to happen exactly once per process.

One test neglected to initialize, and another 
did not do so via the `Once` mechanism. This
was mostly masked by the fact that we run all
these tests in a single process, along with
several other tests that initialized correctly
using `Once`, so correct initialization for that
process was a matter of race condition luck.

However the race is now being lost on the new
macos13 self-hosted runner. And the issue
can be reliably reproduced by running just the
affected test in isolation.

This change ensures that all relevant tests
initialize correctly.