checkpoint: we can make a scie!
Chose to re-use PexBinary target and PexFromTargetsRequest because
there is a lot of existing functionality and asumptions in Pants
around what "a pex" is (and that *a* pex is an artifact and not N
artifcats) and how you can run them. (and that they are runnable) The
tradeoff is that we only support making pex + scie, and not
--scie-only and also that you can't `pants run` the scie pants.  I
think this is a reasonable tradeoff and lets people do what they
want (pass args to pex to make scies) without Pants interjecting some
sort of complicated target generation in their way.

Did need to plumb things down to PexRequest to inject the artifacts.

up next: utility functiosn for calcualting the expectied output/dir
names both for the output_ args and the artifacts