[pants-ng] Initial implementation of contextual config (#22862)
The idea is that instead of global config there can be lots of 
local config files that apply to the directories below them.
The config to apply to a target or file is the union of the config
in its dir, then the config in its parent, and so on up to the
global config in the buildroot. The usual rules for overriding
or merging values across multiple config files will apply, with
the more global config having lower precedence than the
more local config.

Then this idea is taken a step further, by allowing further
contextual filtering based on tags in the filename.
For example, `pants_ci.toml` would only apply in ci, 
`pants_macos.toml` would only apply on macos, 
`pants_ci.macos.toml` would only apply on ci on macos, etc.

This can even support things like `pants_python311.toml`
to have different config for different python interpreters
(these contexts would be requested by the user via some 
as-yet-to-be-determined "context" flag).

The overall goal is to get rid of the duplication of concepts in 
BUILD files and in config in the vast majority of cases. Instead
of having a global option that can be overridden in a special
target field, on a case by case bases, we can often just have config
all the way down (literally!), with Pants doing the partitioning and
invoking the same rule multiple times, each with a different
(subsystems, lists of files) pair.

This change only handles discovery of config files. It does not
wire them up to anything. It is unlikely that existing goal 
implementations can be easily changed to use this mechanism,
so it is probably a pants-ng thing.

The existing config parser continues to be used, but its output
is no longer a global singleton.